Added the date range to the alerts_history()
Added the date range to line 15 along with the adding "dates" to the options_date array and inputs to the request on line 25. To use the date ranges, in the URL request, mode==0, which is the index of dates on the options_date array. That along with "fromDate={}" and "toDate={}" in the URL request in the format "XX.YY.ZZZZ" XX=day, YY=month, ZZZZ=year. ex: "12.05.2021".